requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pandas==2.1.3
numpy==1.25.2
python-multipart==0.0.6
